feat(amb): add specialized time response results and refactor AMB modules
- Created `AmbTimeResponseResults` class in `ross/results.py` to handle specialized plotting for AMB displacement, current, and force responses using Plotly.

- Refactored `TimeResponseAmb` to `AmbTimeResponse` and decoupled utility functions into `ross/bearings/magnetic/amb_utils.py`.

- Moved AMB rotor example factory functions to a dedicated `ross/bearings/magnetic/amb_models.py` file.

- Updated `Rotor.time_response()` to automatically return `AmbTimeResponseResults` when AMBs are present in the assembly.

- Added comprehensive unit tests for values and save/load state functionality in AMB simulations.
