+++ ross/rotor_assembly.py
@@ -27,6 +27,9 @@ from ross.bearing_seal_element import (
     RollerBearingElement,
     SealElement,
 )
+
+from ross.bearings.magnetic.amb_time_response import AmbTimeResponse
+from ross.bearings.magnetic.amb_utils import get_ambs, has_ambs
 from ross.coupling_element import CouplingElement
 from ross.disk_element import DiskElement
 from ross.faults import Crack, MisalignmentFlex, MisalignmentRigid, Rubbing
@@ -48,6 +51,7 @@ from ross.results import (
     TimeResponseResults,
     UCSResults,
     ClearanceResults,
+    AmbTimeResponseResults,
 )
 from ross.seals.labyrinth_seal import LabyrinthSeal
 from ross.shaft_element import ShaftElement
@@ -71,7 +75,8 @@ __all__ = [
     "coaxrotor_example",
     "rotor_example_6dof",
     "rotor_example_with_damping",
-    "rotor_amb_example",
+    "rotor_example_amb_general_controllers",
+    "rotor_example_amb_complex_controllers",
     "concatenate_rotor",
 ]
 
@@ -1936,7 +1941,7 @@ class Rotor(object):
         Examples
         --------
         >>> import ross as rs
-        >>> rotor = rs.rotor_amb_example()
+        >>> rotor = rs.rotor_example_amb_general_controllers()
 
         >>> # Run sensitivity for all magnetic bearings in the rotor (default sweep)
         >>> sensitivity_results = rotor.run_amb_sensitivity(speed=314.16, t_max=5e-4, dt=1e-4) # doctest: +ELLIPSIS
@@ -2433,7 +2438,7 @@ class Rotor(object):
         --------
         >>> import ross as rs
         >>> import numpy as np
-        >>> rotor = rs.rotor_amb_example()
+        >>> rotor = rs.rotor_example_amb_general_controllers()
         >>> dt, speed, step = 1e-4, 1000, 1
         >>> t = np.arange(0, 5 * dt, dt)
         >>> node = [27, 29]
@@ -2850,11 +2855,7 @@ class Rotor(object):
         vectors. Each bearing's controller is rebuilt based on the provided
         time increment.
         """
-        magnetic_bearings = [
-            brg
-            for brg in self.bearing_elements
-            if isinstance(brg, MagneticBearingElement)
-        ]
+        magnetic_bearings = get_ambs(self)
 
         rotor = deepcopy(self)
 
@@ -2932,17 +2933,24 @@ class Rotor(object):
         >>> size = 28
         >>> t = np.linspace(0, 5, size)
         >>> F = np.ones((size, rotor.ndof))
-        >>> rotor.time_response(speed, F, t) # doctest: +ELLIPSIS
-        (array([0.        , 0.18518519, 0.37037037, ...
+        >>> time_response = rotor.time_response(speed, F, t)
+        >>> time_response.yout  # doctest: +ELLIPSIS
+        array([[ 0.00000000e+00,  0.00000000e+00,  0.00000000e+00, ...
         """
 
         if isinstance(speed, Iterable) or method.lower() == "newmark":
             t_, yout = self.integrate_system(speed, F, t, **kwargs)
-            return t_, yout, []
+            return TimeResponseResults(self, t_, yout, [])
+
+        elif has_ambs(self):
+            sim = AmbTimeResponse(self, t=t, speed=speed, F=F, **kwargs)
+            t_, yout, xout = sim.run()
+            return AmbTimeResponseResults(self, t_, yout, xout)
 
         else:
             lti = self._lti(speed)
-            return signal.lsim(lti, F, t, X0=ic)
+            t_, yout, xout = signal.lsim(lti, F, t, X0=ic)
+            return TimeResponseResults(self, t_, yout, xout)
 
     def plot_rotor(self, nodes=1, check_sld=False, length_units="m", **kwargs):
         """Plot a rotor object.
@@ -3614,10 +3622,7 @@ class Rotor(object):
         >>> # plot orbit response - plotting 3D orbits - full rotor model:
         >>> fig3 = response.plot_3d()
         """
-        t_, yout, xout = self.time_response(speed, F, t, method=method, **kwargs)
-        results = TimeResponseResults(self, t, yout, xout)
-
-        return results
+        return self.time_response(speed, F, t, method=method, **kwargs)
 
     @check_units
     def run_harmonic_balance_response(
@@ -5784,165 +5789,16 @@ def rotor_example_with_damping():
     return Rotor(shaft_elem, [disk0, disk1], [bearing0, bearing1])
 
 
-def rotor_amb_example(controller_transfer_function=None):
-    r"""This function creates the model of a test rig rotor supported by magnetic bearings.
-    Details of the model can be found at doi.org/10.14393/ufu.di.2015.186.
+def rotor_example_amb_complex_controllers():
+    from ross.bearings.magnetic.amb_models import rotor_example_amb_complex_controllers
 
-    Returns
-    -------
-    Rotor object.
-    """
+    return rotor_example_amb_complex_controllers()
 
-    from ross.materials import Material
 
-    steel_amb = Material(name="steel", rho=7850, E=1.9e11, Poisson=0.30)
-    steel_m12_amb = Material(name="steel", rho=7600, E=2e11, Poisson=0.31, color="red")
+def rotor_example_amb_general_controllers(controller_transfer_function=None):
+    from ross.bearings.magnetic.amb_models import rotor_example_amb_general_controllers
 
-    # Shaft elements:
-    # fmt: off
-    Li = [
-        0.0, 0.012, 0.032, 0.052, 0.072, 0.092, 0.112, 0.1208,
-        0.12724, 0.13475, 0.14049, 0.14689, 0.15299, 0.159170,
-        0.16535, 0.180350, 0.1905, 0.2063, 0.2221, 0.2379, 0.2537,
-        0.2695, 0.2853, 0.3011, 0.3169, 0.3243, 0.3363, 0.358,
-        0.364, 0.3705, 0.3825, 0.3986, 0.4147, 0.4308, 0.4469,
-        0.4630, 0.4791, 0.4952, 0.5113, 0.5274, 0.5356, 0.5457,
-        0.5607, 0.5669, 0.5731, 0.5792, 0.5856, 0.5913, 0.5989,
-        0.6053, 0.6141, 0.6341, 0.6461,
-    ]
-
-    Li = [round(i, 4) for i in Li]
-    L = [Li[i + 1] - Li[i] for i in range(len(Li) - 1)]
-
-    i_d = [0.0 for i in L]
-    o_d1 = [0.0 for i in L]
-    o_d1[0] = 6.35
-    o_d1[1:5] = [32 for i in range(4)]
-    o_d1[5:14] = [34.8 for i in range(9)]
-    o_d1[14:16] = [49.9 for i in range(2)]
-    o_d1[16:27] = [19.05 for i in range(11)]
-    o_d1[27:29] = [54 for i in range(2)]
-    o_d1[29:40] = [19.05 for i in range(12)]
-    o_d1[40:42] = [49.9 for i in range(2)]
-    o_d1[42:51] = [34.8 for i in range(9)]
-    o_d1[51] = 6.35
-    o_d = [i * 1e-3 for i in o_d1]
-
-    shaft_elements = [
-        ShaftElement(
-            L=l,
-            idl=idl,
-            odl=odl,
-            material=steel_amb,
-            shear_effects=True,
-            rotary_inertia=True,
-            gyroscopic=True,
-            alpha=2.5,
-        )
-        for l, idl, odl in zip(L, i_d, o_d)
-    ]
-
-    # Disk elements:
-    n_list = [27, 28, 29]
-    n_list_2 = [6, 7, 8, 9, 10, 11, 12, 13, 43, 44, 45, 46, 47, 48, 49, 50]
-    width = [0.004, 0.007, 0.014]
-    width_2 = [
-        0.0088, 0.0064, 0.0075, 0.0057, 0.0064, 0.0061,
-        0.0062, 0.0062, 0.0062, 0.0062,0.0061, 0.0064,
-        0.0057, 0.0075, 0.0064, 0.0088,
-    ]
-    i_disc_1 = [0.054, 0.054, 0.054]
-    i_disc_2 = [0.0348] * 16
-    o_disc = [0.1200] * 3
-    o_disc_2 = [0.0498] * 16
-    # fmt: on
-    disk_elements_1 = [
-        DiskElement.from_geometry(n=n, material=steel_amb, width=m, i_d=Id, o_d=Od)
-        for n, m, Id, Od in zip(n_list, width, i_disc_1, o_disc)
-    ]
-
-    disk_elements_2 = [
-        DiskElement.from_geometry(n=n, material=steel_m12_amb, width=m, i_d=Id, o_d=Od)
-        for n, m, Id, Od in zip(n_list_2, width_2, i_disc_2, o_disc_2)
-    ]
-
-    disk_elements = [*disk_elements_1, *disk_elements_2]
-
-    # Bearing elements:
-    n_list = [12, 43]
-    n = 200
-    A = 1e-4
-    i0 = 1.0
-    s0 = 1e-3
-    alpha = 0.392
-    k_amp = 1.0
-    k_sense = 1.0
-
-    if controller_transfer_function is None:
-        Kp = 1000
-        Ki = 0
-        Kd = 5
-
-        bearing_elements = [
-            MagneticBearingElement(
-                n=n_list[0],
-                g0=s0,
-                i0=i0,
-                ag=A,
-                nw=n,
-                alpha=alpha,
-                k_amp=k_amp,
-                k_sense=k_sense,
-                kp_pid=Kp,
-                kd_pid=Kd,
-                ki_pid=Ki,
-                tag="Magnetic Bearing 0",
-            ),
-            MagneticBearingElement(
-                n=n_list[1],
-                g0=s0,
-                i0=i0,
-                ag=A,
-                nw=n,
-                alpha=alpha,
-                k_amp=k_amp,
-                k_sense=k_sense,
-                kp_pid=Kp,
-                kd_pid=Kd,
-                ki_pid=Ki,
-                tag="Magnetic Bearing 1",
-            ),
-        ]
-
-    else:
-        bearing_elements = [
-            MagneticBearingElement(
-                n=n_list[0],
-                g0=s0,
-                i0=i0,
-                ag=A,
-                nw=n,
-                alpha=alpha,
-                k_amp=k_amp,
-                k_sense=k_sense,
-                controller_transfer_function=controller_transfer_function,
-                tag="Magnetic Bearing 0",
-            ),
-            MagneticBearingElement(
-                n=n_list[1],
-                g0=s0,
-                i0=i0,
-                ag=A,
-                nw=n,
-                alpha=alpha,
-                k_amp=k_amp,
-                k_sense=k_sense,
-                controller_transfer_function=controller_transfer_function,
-                tag="Magnetic Bearing 1",
-            ),
-        ]
-
-    return Rotor(shaft_elements, disk_elements, bearing_elements)
+    return rotor_example_amb_general_controllers(controller_transfer_function)
 
 
 def concatenate_rotor(rotor_list):
