+++ ross/stochastic/st_rotor_assembly.py
@@ -635,7 +635,8 @@ class ST_Rotor(object):
         if len(force.shape) == 2:
             # Monte Carlo - results storage
             for i, rotor in enumerate(iter(self)):
-                t_, y, x = rotor.time_response(speed, force, time_range, ic)
+                response = rotor.time_response(speed, force, time_range, ic)
+                t_, y, x = response.t, response.yout, response.xout
                 xout[i] = x
                 yout[i] = y
 
@@ -644,7 +645,8 @@ class ST_Rotor(object):
             # Monte Carlo - results storage
             i = 0
             for rotor, F in zip(iter(self), force):
-                t_, y, x = rotor.time_response(speed, F, time_range, ic)
+                response = rotor.time_response(speed, F, time_range, ic)
+                t_, y, x = response.t, response.yout, response.xout
                 xout[i] = x
                 yout[i] = y
                 i += 1
